speech_micro/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
SpeechRecognition==3.10.0
pydub==0.25.1
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
soundfile==0.12.1
numpy>=1.21.0,<2.0.0
requests==2.31.0
aiofiles==23.2.0
# pyaudio==0.2.11  # Now needed for real-time recording
jose[cryptography]==1.4.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
cryptography==41.0.8
google-auth==2.25.2
google-auth-transport-requests==0.1.3
websockets==12.0